+++ server/entry-node.ts
@@ -13,7 +13,7 @@ import { createNodePlatform } from './platform/node'
 import { type DeployPlatform, setDeployPlatform } from './runtime-platform'
 import { syncPendingCloudTrafficReports } from './usecases/cloud-traffic-metering'
 import { syncPendingRemoteDownloadUsageReports } from './usecases/downloads/remote-download-usage'
-import { purgeExpiredTrash, resolveTrashRetentionDays } from './usecases/purge'
+import { purgeExpiredTrash, resolveTrashRetentionDays } from './usecases/object'
 import { buildCloudInstanceInfo, runtimeInfo } from './usecases/site/instance-info'
 import { INSTANCE_TELEMETRY_CRON, reportInstanceTelemetry } from './usecases/site/instance-telemetry'
 import { runLicensingRefresh } from './usecases/site/licensing'
