+++ server/usecases/object.ts
@@ -26,25 +26,23 @@ import { meterDownloadTraffic } from './cloud-traffic-metering'
 import type { Deps } from './deps'
 import { assertTaskUploadAllowed } from './downloads/downloads'
 import {
-  createObjectUploadSession,
+  type ActivityRepo,
+  type Matter,
+  type MatterListFilters,
+  type MatterRepo,
   ObjectUploadSessionError,
-  patchObjectUploadSession,
-  presignObjectUploadParts,
-} from './object-upload-session'
-import type {
-  ActivityRepo,
-  Matter,
-  MatterListFilters,
-  MatterRepo,
-  QuotaRepo,
-  StorageRecord,
-  StorageUsageRepo,
+  type ObjectUploadSessionRecord,
+  type ObjectUploadSessionRepo,
+  type QuotaRepo,
+  type S3Gateway,
+  type ShareRepo,
+  type StorageRecord,
+  type StorageRepo,
+  type StorageUsageRepo,
 } from './ports'
-import { purgeRecursively } from './purge'
-import { copyMatterToOrg } from './save-to-drive'
 import { StorageQuotaExceededError, withStorageUsageReservation } from './storage-usage'
 
-export { ObjectUploadSessionError } from './object-upload-session'
+export { ObjectUploadSessionError } from './ports'
 
 // The copy request shape (mirrors copyMatterSchema; no shared type is exported).
 export interface CopyObjectInput {
@@ -662,4 +660,441 @@ export async function confirmUpload(
     if (error instanceof Error && error.message === 'CONFIRM_UPLOAD_RACE') return { matter: null }
     throw error
   }
+}
+
+// ── upload sessions ──────────────────────────────────────────────────────────
+
+export type ObjectUploadSessionDeps = { s3: S3Gateway; objectUploadSessions: ObjectUploadSessionRepo }
+
+const DEFAULT_PART_SIZE = 16 * 1024 * 1024
+
+function toDto(record: ObjectUploadSessionRecord): ObjectUploadSession {
+  return {
+    id: record.id,
+    objectId: record.objectId,
+    uploadId: record.uploadId,
+    partSize: record.partSize,
+    status: record.status,
+    expiresAt: record.expiresAt.toISOString(),
+    createdAt: record.createdAt.toISOString(),
+    updatedAt: record.updatedAt.toISOString(),
+  }
+}
+
+export async function createObjectUploadSession(
+  deps: ObjectUploadSessionDeps,
+  params: {
+    orgId: string
+    objectId: string
+    storage: StorageRecord
+    storageKey: string
+    contentType: string
+    partSize?: number
+    actorId: string
+  },
+): Promise<ObjectUploadSession> {
+  let uploadId: string
+  try {
+    uploadId = await deps.s3.createMultipartUpload(params.storage, params.storageKey, params.contentType)
+  } catch (error) {
+    throw new ObjectUploadSessionError(
+      'storage_failure',
+      `Storage multipart upload failed: ${(error as Error).message}`,
+    )
+  }
+  const record = await deps.objectUploadSessions.create({
+    orgId: params.orgId,
+    objectId: params.objectId,
+    storageId: params.storage.id,
+    storageKey: params.storageKey,
+    uploadId,
+    partSize: params.partSize ?? DEFAULT_PART_SIZE,
+    actorId: params.actorId,
+  })
+  return toDto(record)
+}
+
+export async function getObjectUploadSession(
+  deps: ObjectUploadSessionDeps,
+  orgId: string,
+  objectId: string,
+  id: string,
+): Promise<ObjectUploadSession> {
+  const record = await deps.objectUploadSessions.get(orgId, objectId, id)
+  if (!record) throw new ObjectUploadSessionError('not_found')
+  return toDto(record)
+}
+
+export async function presignObjectUploadParts(
+  deps: ObjectUploadSessionDeps,
+  params: {
+    orgId: string
+    objectId: string
+    sessionId: string
+    storage: StorageRecord
+    partNumbers: number[]
+  },
+): Promise<{ uploadId: string; partSize: number; parts: Array<{ partNumber: number; url: string }> }> {
+  const record = await deps.objectUploadSessions.get(params.orgId, params.objectId, params.sessionId)
+  if (!record) throw new ObjectUploadSessionError('not_found')
+  if (record.status !== 'active' || record.expiresAt.getTime() <= Date.now()) {
+    throw new ObjectUploadSessionError('invalid_state')
+  }
+  const parts = await Promise.all(
+    params.partNumbers.map(async (partNumber) => ({
+      partNumber,
+      url: await deps.s3.presignUploadPart(params.storage, record.storageKey, record.uploadId, partNumber),
+    })),
+  )
+  return { uploadId: record.uploadId, partSize: record.partSize, parts }
+}
+
+export async function patchObjectUploadSession(
+  deps: ObjectUploadSessionDeps,
+  params: {
+    orgId: string
+    objectId: string
+    sessionId: string
+    storage: StorageRecord
+    input: PatchObjectUploadSessionInput
+  },
+): Promise<ObjectUploadSession> {
+  const record = await deps.objectUploadSessions.get(params.orgId, params.objectId, params.sessionId)
+  if (!record) throw new ObjectUploadSessionError('not_found')
+  if (record.status !== 'active') throw new ObjectUploadSessionError('invalid_state')
+  if (params.input.action === 'complete') {
+    try {
+      await deps.s3.completeMultipartUpload(params.storage, record.storageKey, record.uploadId, params.input.parts)
+    } catch (error) {
+      throw new ObjectUploadSessionError(
+        'storage_failure',
+        `Storage multipart upload complete failed: ${(error as Error).message}`,
+      )
+    }
+    await deps.objectUploadSessions.setStatus(record.id, 'completed')
+  } else {
+    try {
+      await deps.s3.abortMultipartUpload(params.storage, record.storageKey, record.uploadId)
+    } catch (error) {
+      throw new ObjectUploadSessionError(
+        'storage_failure',
+        `Storage multipart upload abort failed: ${(error as Error).message}`,
+      )
+    }
+    await deps.objectUploadSessions.setStatus(record.id, 'aborted')
+  }
+  return getObjectUploadSession(deps, params.orgId, params.objectId, params.sessionId)
+}
+
+// ── recursive purge ──────────────────────────────────────────────────────────
+
+const DAY_MS = 24 * 60 * 60 * 1000
+export const DEFAULT_TRASH_RETENTION_DAYS = 30
+
+export type PurgeDeps = {
+  s3: S3Gateway
+  storages: StorageRepo
+  storageUsage: StorageUsageRepo
+  share: ShareRepo
+  matter: MatterRepo
+}
+
+export async function purgeRecursively(deps: PurgeDeps, orgId: string, matters: Matter[]): Promise<number> {
+  const keysByStorage = new Map<string, { storage: StorageRecord | null; keys: string[] }>()
+  const bytesByStorage = new Map<string, number>()
+  let totalBytes = 0
+
+  for (const m of matters) {
+    const size = m.size ?? 0
+    if (m.dirtype === DirType.FILE && size > 0) {
+      bytesByStorage.set(m.storageId, (bytesByStorage.get(m.storageId) ?? 0) + size)
+      totalBytes += size
+    }
+    if (!m.object) continue
+    let entry = keysByStorage.get(m.storageId)
+    if (!entry) {
+      const storage = await deps.storages.get(m.storageId)
+      entry = { storage, keys: [] }
+      keysByStorage.set(m.storageId, entry)
+    }
+    entry.keys.push(m.object)
+  }
+
+  for (const { storage, keys } of keysByStorage.values()) {
+    if (storage && keys.length > 0) await deps.s3.deleteObjects(storage, keys)
+  }
+
+  for (const m of matters) {
+    await deps.share.cascadeDeleteByMatter(m.id)
+  }
+
+  await deps.matter.purge(
+    orgId,
+    matters.map((m) => m.id),
+  )
+  if (totalBytes > 0) await deps.storageUsage.reconcile(orgId, bytesByStorage.keys())
+  return matters.length
+}
+
+/** Parses ZPAN_TRASH_RETENTION_DAYS; falls back to the default, 0 disables purge. */
+export function resolveTrashRetentionDays(raw: string | undefined): number {
+  if (raw === undefined || raw.trim() === '') return DEFAULT_TRASH_RETENTION_DAYS
+  const days = Number(raw)
+  if (!Number.isFinite(days) || days < 0) return DEFAULT_TRASH_RETENTION_DAYS
+  return Math.floor(days)
+}
+
+/**
+ * Permanently purges trashed items older than `retentionDays` across all orgs,
+ * reclaiming their quota. Retention of 0 disables auto-purge. Runs subtree at a
+ * time via the same purge path as emptying the trash manually.
+ */
+export async function purgeExpiredTrash(deps: PurgeDeps, retentionDays: number, now = Date.now()): Promise<number> {
+  if (retentionDays <= 0) return 0
+  const cutoff = now - retentionDays * DAY_MS
+  const orgIds = await deps.matter.listOrgIdsWithExpiredTrash(cutoff)
+
+  let purged = 0
+  for (const orgId of orgIds) {
+    const roots = await deps.matter.listTrashedRoots(orgId)
+    for (const root of roots) {
+      if ((root.trashedAt ?? 0) >= cutoff) continue
+      const matters = await deps.matter.collectForPurge(orgId, root.id)
+      if (!matters) continue
+      purged += await purgeRecursively(deps, orgId, matters)
+    }
+  }
+  return purged
+}
+
+// ── save to drive ────────────────────────────────────────────────────────────
+
+// Pure orchestration: copies a shared matter (file or folder) into another org,
+// reserving target-org quota per file via withStorageUsageReservation. Reaches
+// the outside world only through deps; matter creation goes through the matter
+// repo port.
+
+export type SaveToDriveDeps = {
+  s3: S3Gateway
+  storages: StorageRepo
+  storageUsage: StorageUsageRepo
+  quota: QuotaRepo
+  activity: ActivityRepo
+  share: ShareRepo
+  matter: MatterRepo
+}
+
+export interface SaveShareInput {
+  share: { id: string }
+  matter: Matter
+  currentUserId: string
+  targetOrgId: string
+  targetParent: string
+  teamQuotaEnabled?: boolean
+}
+
+export interface SaveShareResult {
+  saved: Matter[]
+  skipped: Array<{ name: string; reason: string }>
+}
+
+interface CopyActivity {
+  action: string
+  metadata: Record<string, unknown>
+}
+
+export interface CopyMatterToOrgInput {
+  sourceMatter: Matter
+  currentUserId: string
+  targetOrgId: string
+  targetParent: string
+  activity: CopyActivity
+  teamQuotaEnabled?: boolean
+}
+
+function buildPath(parent: string, name: string): string {
+  return parent ? `${parent}/${name}` : name
+}
+
+async function saveFile(
+  deps: SaveToDriveDeps,
+  sourceMatter: Matter,
+  sourceStorage: StorageRecord,
+  targetStorage: StorageRecord,
+  currentUserId: string,
+  targetOrgId: string,
+  targetParent: string,
+  activity: CopyActivity,
+  teamQuotaEnabled = true,
+): Promise<Matter> {
+  const bytes = sourceMatter.size ?? 0
+  const dstKey = buildObjectKey({ uid: currentUserId, orgId: targetOrgId, rawExt: fileExt(sourceMatter.name) })
+
+  return withStorageUsageReservation(
+    { quota: deps.quota, storageUsage: deps.storageUsage },
+    { orgId: targetOrgId, storageId: targetStorage.id, bytes, teamQuotaEnabled },
+    async (ctx) => {
+      if (sourceStorage.id === targetStorage.id) {
+        await deps.s3.copyObject(sourceStorage, sourceMatter.object, targetStorage, dstKey)
+      } else {
+        await deps.s3.streamCopy(sourceStorage, sourceMatter.object, targetStorage, dstKey)
+      }
+      ctx.onRollback(() => deps.s3.deleteObject(targetStorage, dstKey))
+
+      const newMatter = await deps.matter.create({
+        orgId: targetOrgId,
+        name: sourceMatter.name,
+        type: sourceMatter.type,
+        size: bytes,
+        dirtype: DirType.FILE,
+        parent: targetParent,
+        object: dstKey,
+        storageId: targetStorage.id,
+        status: 'active',
+        onConflict: 'rename',
+      })
+
+      await deps.activity.record({
+        orgId: targetOrgId,
+        userId: currentUserId,
+        action: activity.action,
+        targetType: 'file',
+        targetId: newMatter.id,
+        targetName: newMatter.name,
+        metadata: activity.metadata,
+      })
+
+      return newMatter
+    },
+  )
+}
+
+async function saveFolderRecursive(
+  deps: SaveToDriveDeps,
+  sourceFolderMatter: Matter,
+  sourceStorage: StorageRecord,
+  targetStorage: StorageRecord,
+  currentUserId: string,
+  targetOrgId: string,
+  targetParent: string,
+  activity: CopyActivity,
+  teamQuotaEnabled = true,
+): Promise<SaveShareResult> {
+  const saved: Matter[] = []
+  const skipped: Array<{ name: string; reason: string }> = []
+
+  const rootFolder = await deps.matter.create({
+    orgId: targetOrgId,
+    name: sourceFolderMatter.name,
+    type: 'folder',
+    size: 0,
+    dirtype: sourceFolderMatter.dirtype ?? undefined,
+    parent: targetParent,
+    object: '',
+    storageId: targetStorage.id,
+    status: 'active',
+    onConflict: 'rename',
+  })
+  saved.push(rootFolder)
+
+  const sourceRootPath = buildPath(sourceFolderMatter.parent, sourceFolderMatter.name)
+  const targetRootPath = buildPath(targetParent, rootFolder.name)
+
+  const queue: Array<{ sourcePath: string; targetPath: string }> = [
+    { sourcePath: sourceRootPath, targetPath: targetRootPath },
+  ]
+
+  while (queue.length > 0) {
+    const { sourcePath, targetPath } = queue.shift()!
+    const children = await deps.share.listDirectActiveChildren(sourceFolderMatter.orgId, sourcePath)
+
+    for (const child of children) {
+      if (child.dirtype === DirType.FILE) {
+        try {
+          const newFile = await saveFile(
+            deps,
+            child,
+            sourceStorage,
+            targetStorage,
+            currentUserId,
+            targetOrgId,
+            targetPath,
+            activity,
+            teamQuotaEnabled,
+          )
+          saved.push(newFile)
+        } catch (e) {
+          skipped.push({ name: child.name, reason: (e as Error).message })
+        }
+      } else {
+        const newFolder = await deps.matter.create({
+          orgId: targetOrgId,
+          name: child.name,
+          type: 'folder',
+          size: 0,
+          dirtype: child.dirtype ?? undefined,
+          parent: targetPath,
+          object: '',
+          storageId: targetStorage.id,
+          status: 'active',
+          onConflict: 'rename',
+        })
+        saved.push(newFolder)
+        queue.push({
+          sourcePath: buildPath(child.parent, child.name),
+          targetPath: buildPath(targetPath, newFolder.name),
+        })
+      }
+    }
+  }
+
+  return { saved, skipped }
+}
+
+// Copy a file or folder (recursively) into another org. Quota is reserved in the
+// target org per file; files that fail (e.g. quota) are reported in `skipped`
+// rather than failing the whole operation.
+export async function copyMatterToOrg(deps: SaveToDriveDeps, input: CopyMatterToOrgInput): Promise<SaveShareResult> {
+  const { sourceMatter, currentUserId, targetOrgId, targetParent, activity, teamQuotaEnabled = true } = input
+
+  const sourceStorage = await deps.storages.get(sourceMatter.storageId)
+  if (!sourceStorage) throw new Error('Source storage not found')
+
+  const targetStorage = await deps.storages.select('private')
+
+  if (sourceMatter.dirtype === DirType.FILE) {
+    const newMatter = await saveFile(
+      deps,
+      sourceMatter,
+      sourceStorage,
+      targetStorage,
+      currentUserId,
+      targetOrgId,
+      targetParent,
+      activity,
+      teamQuotaEnabled,
+    )
+    return { saved: [newMatter], skipped: [] }
+  }
+
+  return saveFolderRecursive(
+    deps,
+    sourceMatter,
+    sourceStorage,
+    targetStorage,
+    currentUserId,
+    targetOrgId,
+    targetParent,
+    activity,
+    teamQuotaEnabled,
+  )
+}
+
+export async function saveShareToDrive(deps: SaveToDriveDeps, input: SaveShareInput): Promise<SaveShareResult> {
+  const { share, matter: sourceMatter, ...rest } = input
+  return copyMatterToOrg(deps, {
+    ...rest,
+    sourceMatter,
+    activity: { action: 'save_from_share', metadata: { sourceShareId: share.id } },
+  })
 }
