+++ workers/scheduled.ts
@@ -5,7 +5,7 @@ import { createDeps } from '../server/composition'
 import { createCloudflarePlatform } from '../server/platform/cloudflare'
 import { syncPendingCloudTrafficReports } from '../server/usecases/cloud-traffic-metering'
 import { syncPendingRemoteDownloadUsageReports } from '../server/usecases/downloads/remote-download-usage'
-import { purgeExpiredTrash, resolveTrashRetentionDays } from '../server/usecases/purge'
+import { purgeExpiredTrash, resolveTrashRetentionDays } from '../server/usecases/object'
 import { INSTANCE_TELEMETRY_CRON, reportInstanceTelemetry } from '../server/usecases/site/instance-telemetry'
 import { runLicensingRefresh } from '../server/usecases/site/licensing'
 import { ZPAN_CLOUD_URL_DEFAULT } from '../shared/constants'
